Configure user based auth for deployments

+++ app.py
@@ -17,8 +17,12 @@ def get_query_params():
     return app.current_request.query_params or {}
 
 
-def get_region():
-    return get_query_params().get("region", AWS_DEFAULT_REGION)
+def get_headers():
+    return app.current_request.headers or {}
+
+
+def get_region_name():
+    return get_query_params().get("region_name", AWS_DEFAULT_REGION)
 
 
 def get_namespace():
@@ -28,9 +32,20 @@ def get_namespace():
 @app.route("/{cluster_name}", methods=["POST"])
 def deploy_pods(cluster_name):
     try:
-        region_name = get_region()
+        aws_access_key_id = get_headers().get("AWS_PUBLIC_KEY")
+        aws_secret_access_key = get_headers().get("AWS_SECRET_KEY")
+
+        if not aws_access_key_id or not aws_secret_access_key:
+            raise UnauthorizedError("AWS_PUBLIC_KEY and AWS_SECRET_KEY are required")
+
+        region_name = get_region_name()
         namespace = get_namespace()
-        kubernetes_client = get_kubernetes_client(cluster_name, region_name=region_name)
+        kubernetes_client = get_kubernetes_client(
+            cluster_name,
+            region_name=region_name,
+            aws_public_key=aws_access_key_id,
+            aws_secret_access_key=aws_secret_access_key,
+        )
         create_deployment(
             kubernetes_client,
             configuration_files=CLUSTER_CONFIGURATION_FILES,
@@ -45,9 +60,20 @@ def deploy_pods(cluster_name):
 @app.route("/{cluster_name}", methods=["DELETE"])
 def destroy_deployed_pods(cluster_name):
     try:
-        region_name = get_region()
+        aws_access_key_id = get_headers().get("AWS_PUBLIC_KEY")
+        aws_secret_access_key = get_headers().get("AWS_SECRET_KEY")
+
+        if not aws_access_key_id or not aws_secret_access_key:
+            raise UnauthorizedError("AWS_PUBLIC_KEY and AWS_SECRET_KEY are required")
+
+        region_name = get_region_name()
         namespace = get_namespace()
-        kubernetes_client = get_kubernetes_client(cluster_name, region_name=region_name)
+        kubernetes_client = get_kubernetes_client(
+            cluster_name,
+            region_name=region_name,
+            aws_access_key_id=aws_access_key_id,
+            aws_secret_access_key=aws_secret_access_key,
+        )
         destroy_deployment(kubernetes_client, namespace=namespace)
 
         return "Destroyed"
